triangle list rendering of cube

+++ MojFramework/Graphics.cpp
@@ -414,8 +414,8 @@ void Graphics::DrawFlatTopTriangle(const Vec2& v0, const Vec2& v1, const Vec2& v
 	float m1 = (v2.x - v1.x) / (v2.y - v1.y);
 
 	// calculate start and end scanlines
-	const int yStart = (int)ceilf(v0.y - 0.5f);
-	const int yEnd = (int)ceilf(v2.y - 0.5f); // the scanline AFTER the last line drawn
+	const int yStart = (int)ceil(v0.y - 0.5f);
+	const int yEnd = (int)ceil(v2.y - 0.5f); // the scanline AFTER the last line drawn
 
 	for (int y = yStart; y < yEnd; y++)
 	{
@@ -425,8 +425,8 @@ void Graphics::DrawFlatTopTriangle(const Vec2& v0, const Vec2& v1, const Vec2& v
 		const float px1 = m1 * (float(y) + 0.5f - v1.y) + v1.x;
 
 		// calculate start and end pixels
-		const int xStart = (int)ceilf(px0 - 0.5f);
-		const int xEnd = (int)ceilf(px1 - 0.5f); // the pixel AFTER the last pixel drawn
+		const int xStart = (int)ceil(px0 - 0.5f);
+		const int xEnd = (int)ceil(px1 - 0.5f); // the pixel AFTER the last pixel drawn
 
 		for (int x = xStart; x < xEnd; x++)
 		{
@@ -442,8 +442,8 @@ void Graphics::DrawFlatBottomTriangle(const Vec2& v0, const Vec2& v1, const Vec2
 	float m1 = (v2.x - v0.x) / (v2.y - v0.y);
 
 	// calculate start and end scanlines
-	const int yStart = (int)ceilf(v0.y - 0.5f);
-	const int yEnd = (int)ceilf(v2.y - 0.5f); // the scanline AFTER the last line drawn
+	const int yStart = (int)ceil(v0.y - 0.5f);
+	const int yEnd = (int)ceil(v2.y - 0.5f); // the scanline AFTER the last line drawn
 
 	for (int y = yStart; y < yEnd; y++)
 	{
@@ -453,8 +453,8 @@ void Graphics::DrawFlatBottomTriangle(const Vec2& v0, const Vec2& v1, const Vec2
 		const float px1 = m1 * (float(y) + 0.5f - v0.y) + v0.x;
 
 		// calculate start and end pixels
-		const int xStart = (int)ceilf(px0 - 0.5f);
-		const int xEnd = (int)ceilf(px1 - 0.5f); // the pixel AFTER the last pixel drawn
+		const int xStart = (int)ceil(px0 - 0.5f);
+		const int xEnd = (int)ceil(px1 - 0.5f); // the pixel AFTER the last pixel drawn
 
 		for (int x = xStart; x < xEnd; x++)
 		{
